+++ app/src/main/java/com/scottyab/rootbeer/sample/extensions/ViewExtensions.kt
@@ -1,0 +1,23 @@
+package com.scottyab.rootbeer.sample.extensions
+
+import android.view.View
+
+internal fun View.show(show: Boolean) {
+    if (show) {
+        show()
+    } else {
+        hide()
+    }
+}
+
+internal fun View.show() {
+    visibility = View.VISIBLE
+}
+
+internal fun View.invisible() {
+    visibility = View.INVISIBLE
+}
+
+internal fun View.hide() {
+    visibility = View.GONE
+}